refactor(api): migrate GarAnonymizedBacthEventLogging to context
