+++ api/src/identity-access-management/domain/models/GarAuthenticationMethodAnonymized.js
